[#18] typesafe schema validation
# Problem Statement
Since the last refactor of acceptance of input schema format, the schema validation is performed through spray-json that parses the schema file and tries to map it in a `ChromoSchema`.

If an invalid schema is detected, a runtime exception is thrown.

# Solution
Schema parsing should return `Either[ChromoSchemaException, ChromoSchema]`.

The separation between a well-formed JSON and a valid JSON needs to be discussed.
Now, both the responsibilities are embedded in the parsing logic: if the JSON is well written AND the chromosoma semantics is valid, then the schema is materialized in chromosoma data structure.

Two responsibilities should be refactored into two entities:

- A `SchemaParser` should materialize the well-formed json schema in staging data structure (in a typesafe fashion)
- A `SchemaValidator` should check the chromosoma semantics, and map the staging data structure to business objects.

Using [circe](https://circe.github.io/circe/) we gain almost free mapping between the schema file and the staging data structure. Then the validation logic needs to be implemented.